@@ -11,27 +11,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
+    "@angular/animations": "^8.0.3",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
+    "@angular/common": "^8.0.3",
+    "@angular/compiler": "~8.0.3",
+    "@angular/core": "^8.0.3",
+    "@angular/forms": "~8.0.3",
     "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
-    "@ng-bootstrap/ng-bootstrap": "^4.2.1",
+    "@angular/platform-browser": "~8.0.3",
+    "@angular/platform-browser-dynamic": "~8.0.3",
+    "@angular/router": "~8.0.3",
+    "@ng-bootstrap/ng-bootstrap": "^5.0.0-rc.1",
     "angular-font-awesome": "^3.1.2",
     "angular-highlight-js": "^2.0.1",
-    "angular-l10n": "^7.2.0",
+    "angular-l10n": "^8.0.0",
     "bootstrap": "^4.3.1",
-    "core-js": "^2.5.4",
+    "core-js": "^3.1.4",
     "font-awesome": "^4.7.0",
     "highlight.js": "^9.15.8",
     "jquery": "^3.4.1",
     "libfhqcli-web-js": "^0.2.22",
-    "marked": "^0.6.2",
+    "marked": "^0.6.3",
     "ngx-ace-wrapper": "^8.0.0",
     "ngx-markdown-editor": "^1.2.0",
     "ngx-toastr": "^10.0.4",
@@ -40,27 +40,27 @@
     "rxjs": "^6.5.2",
     "rxjs-compat": "^6.5.2",
     "tslib": "^1.10.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.9",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~2.8.8",
+    "@angular-devkit/build-angular": "~0.800.6",
+    "@angular/cli": "~8.0.6",
+    "@angular/compiler-cli": "~8.0.3",
+    "@angular/language-service": "~8.0.3",
+    "@types/jasmine": "~3.3.13",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "@types/node": "~12.0.10",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "ts-node": "~8.3.0",
+    "tslint": "~5.18.0",
+    "typescript": "3.4.5"
   }
 }
